chore: split ast functions into it's own file
