+++ src/lib/index.d.ts
@@ -1,0 +1,2 @@
+declare module '@babel/preset-typescript'
+declare module '@babel/generator'
